refactor(cli): rename lib/api/ to lib/{server,environment}
After the platform-client wrapper was deleted, `lib/api/` held only two
unrelated CLI-side helpers: `resolve-server.ts` (picks which Zitadel
URL to talk to) and `schemas.ts` (a 9-line zod enum for the
`--environment` flag). The directory name was misleading — neither has
anything to do with the API client.

- `lib/api/resolve-server.ts` → `lib/server.ts`
- `lib/api/schemas.ts` → `lib/environment.ts`
- `lib/api/index.ts` deleted (was just two re-exports)
- `lib/api/` directory removed

Imports updated across `commands/{plan,apply,setup}.ts`,
`commands/setup/prompts/server.ts`, `lib/oclif/base.ts`, and
`lib/orca/patchers/rule/base.ts`. Test files moved to the mirrored
paths and their import depth fixed.

Also drops a redundant `setApiAuthToken(undefined)` from `setup` —
the registry defaults to `undefined` in a fresh process, so the
defensive clear was noise.

`environment.ts`'s docblock had a dangling reference to the now-deleted
`api/client.ts`; replaced with a generic "see @zitadel-nextgen/api".
