+++ apps/cli/src/lib/orca/patchers/rule/base.ts
@@ -1,6 +1,6 @@
 import { buildFlow } from "../../../flows";
 import { stableStringify } from "../../../json";
-import { DEFAULT_SERVER } from "../../../api/resolve-server";
+import { DEFAULT_SERVER } from "../../../server";
 import { scaffold } from "./file-writer";
 import type { FileOp, ScaffoldPlan } from "./file-writer/types";
 import type {
